- Added BinaryObjectInfo project (refactored & improved version of code in old ConvertObjectBinary). - Added ConvertBinaryObject project (refactored & improved version of code in old ConvertObjectBinary). - Implemented conditional inherit/merge in binary class definitions. - Added ComputeHash32 and ComputeHash64 field types. - Cleaned up small issues in FileFormats, ConvertXml, Pack, Unpack.

+++ Gibbed.Dunia2.FileFormats/Big/EntryCompression.cs
@@ -81,7 +81,7 @@ namespace Gibbed.Dunia2.FileFormats.Big
                 entry.CompressionScheme = CompressionScheme.LZO1x;
                 entry.UncompressedSize = uncompressedSize;
                 entry.CompressedSize = (uint)actualCompressedSize;
-                output.Write(compressedData, 0, (int)actualCompressedSize);
+                output.Write(compressedData, 0, actualCompressedSize);
             }
             else
             {
